% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_fourier.R
\name{coef_fourier}
\alias{coef_fourier}
\title{This function computes minimum square estimates for Fourier coefficients.}
\usage{
coef_fourier(f, m)
}
\arguments{
\item{f}{A time series to be smoothed.}

\item{m}{Order of the Fourier polynomial. Default value is computed using
the Sturge's rule.}
}
\value{
A vector with the fourier coefficients.
}
\description{
This function computes minimum square estimates for Fourier coefficients.
}
\examples{
data(canada)

coef_fourier(canada$ThePas_ts)
}
