% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_b.R
\name{fourier_b}
\alias{fourier_b}
\title{This function the smoothed curve}
\usage{
fourier_b(coef, x)
}
\arguments{
\item{coef}{Fourier coefficients.}

\item{x}{a time series to evaluate the smoothed curve.}
}
\value{
a time series with the smoothed curve.
}
\description{
This function the smoothed curve
}
\examples{
data(canada)

coefs <- coef_fourier(canada$ThePas_ts)
y_hat <- fourier_b(coefs)
}
