% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLik}
\alias{logLik}
\title{Log-likehood function multiplied by -1.}
\arguments{
\item{mDist}{distance matris;}

\item{s2}{variance from the covariance model;}

\item{phi}{variance from the covariance model;}

\item{vDiff}{column vector of data (subtracted the mean vector)}
}
\value{
log-likelihood value multiplied by -1.
}
\description{
This function computes the likelihood function
used at \code{geo_model}.
}
