% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLik}
\alias{logLik}
\title{Log likelihood function for normal distribution with spatial
dependency.}
\arguments{
\item{mDist}{distance matris;}

\item{s2}{variance from the covariance model;}

\item{phi}{variance from the covariance model;}

\item{vDiff}{column vector of data (subtracted the mean vector)}
}
\description{
Log likelihood function for normal distribution with spatial
dependency.
}
